+++ airslate/client.py
@@ -13,7 +13,7 @@ import requests
 from asdicts.dict import merge, intersect_keys
 
 from . import exceptions, session
-from .flow import Flow
+from .facades import Flow, Slate
 from .resources.addons import Addons
 from .resources.slate_addon import SlateAddonFiles
 from .utils import default_headers
@@ -86,6 +86,7 @@ class Client:
         # Initialize each resource and injecting client object into it
         self.addons = Addons(self)
         self.flow = Flow(self)
+        self.slate = Slate(self)
         self.slate_addon_files = SlateAddonFiles(self)
 
     def request(self, method: str, path: str, **options):
